+++ modules/item/race-item-sheet.js
@@ -1,0 +1,13 @@
+import { HMItemSheet } from './item-sheet.js';
+
+export class HMRaceItemSheet extends HMItemSheet {
+    /** @override */
+    static get defaultOptions() {
+        return mergeObject(super.defaultOptions, {
+            classes: ['hackmaster', 'sheet', 'item'],
+            width: 600,
+            height: 400,
+            tabs: [{ navSelector: '.sheet-tabs', contentSelector: '.sheet-body', initial: 'description' }],
+        });
+    }
+}